+++ test/database_test.rb
@@ -1,6 +1,6 @@
 require_relative "test_helper"
 
-class ActiveRecordTest < Minitest::Test
+class DatabaseTest < Minitest::Test
   def setup
     User.delete_all
   end
@@ -23,36 +23,36 @@ class ActiveRecordTest < Minitest::Test
     assert_equal Polars::Int64, df["number"].dtype
   end
 
-  def test_read_sql_relation
+  def test_read_database_relation
     users = create_users
-    df = Polars.read_sql(User.order(:id))
+    df = Polars.read_database(User.order(:id))
     assert_equal ["id", "name", "number"], df.columns
     assert_series users.map(&:id), df["id"]
     assert_series users.map(&:name), df["name"]
     assert_equal Polars::Int64, df["number"].dtype
   end
 
-  def test_read_sql_result
+  def test_read_database_result
     users = create_users
-    df = Polars.read_sql(User.connection.select_all("SELECT * FROM users ORDER BY id"))
+    df = Polars.read_database(User.connection.select_all("SELECT * FROM users ORDER BY id"))
     assert_equal ["id", "name", "number"], df.columns
     assert_series users.map(&:id), df["id"]
     assert_series users.map(&:name), df["name"]
     assert_equal Polars::Int64, df["number"].dtype
   end
 
-  def test_read_sql_string
+  def test_read_database_string
     users = create_users
-    df = Polars.read_sql("SELECT * FROM users ORDER BY id")
+    df = Polars.read_database("SELECT * FROM users ORDER BY id")
     assert_equal ["id", "name", "number"], df.columns
     assert_series users.map(&:id), df["id"]
     assert_series users.map(&:name), df["name"]
     assert_equal Polars::Int64, df["number"].dtype
   end
 
-  def test_read_sql_unsupported
+  def test_read_database_unsupported
     error = assert_raises(ArgumentError) do
-      Polars.read_sql(Object.new)
+      Polars.read_database(Object.new)
     end
     assert_equal "Expected ActiveRecord::Relation, ActiveRecord::Result, or String", error.message
   end